requirements.txt
pillow
jinja2
python-dotenv
